# docs/api/radiance-openapi.json
{
  "components": {
    "schemas": {
...
      "PublicErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "const": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error",
          "message",
          "correlation_id"
        ],
        "title": "PublicErrorResponse",
        "type": "object"
      },
...
      "RadianceManifestResponse": {
        "additionalProperties": false,
        "properties": {
          "manifest": {
            "title": "Manifest",
            "type": "object"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "manifest",
          "path"
        ],
        "title": "RadianceManifestResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Horticulture Lighting Simulator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/radiance/image": {
      "get": {
        "operationId": "get_radiance_image",
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "required": true,
            "schema": {
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "execution_mode",
            "required": false,
            "schema": {
              "default": "precomputed",
              "title": "Execution Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sim_mode",
            "required": false,
            "schema": {
              "default": "standard",
              "title": "Sim Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_ppfd",
            "required": false,
            "schema": {
              "default": 1000.0,
              "title": "Target Ppfd",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "peak_capping_enabled",
            "required": false,
            "schema": {
              "default": false,
              "title": "Peak Capping Enabled",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "match_system_ppe",
            "required": false,
            "schema": {
              "default": false,
              "title": "Match System Ppe",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "length_ft",
            "required": false,
            "schema": {
              "default": 10.0,
              "title": "Length Ft",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "width_ft",
            "required": false,
            "schema": {
              "default": 10.0,
              "title": "Width Ft",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "w_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "W Min"
            }
          },
          {
            "in": "query",
            "name": "w_max",
            "required": false,
            "schema": {
              "default": 100.0,
              "title": "W Max",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "competitor_layout",
            "required": false,
            "schema": {
              "default": "full",
              "title": "Competitor Layout",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hps_coverage_ft",
            "required": false,
            "schema": {
              "default": 4.0,
              "title": "Hps Coverage Ft",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "hps_ies_variant",
            "required": false,
            "schema": {
              "default": "karma",
              "title": "Hps Ies Variant",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mount_z_m",
            "required": false,
            "schema": {
              "default": 0.4572,
              "title": "Mount Z M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "sp_z_m",
            "required": false,
            "schema": {
              "default": 0.4572,
              "title": "Sp Z M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "hps_z_m",
            "required": false,
            "schema": {
              "default": 0.4572,
              "title": "Hps Z M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "basis_backend",
            "required": false,
            "schema": {
              "default": "rtrace",
              "title": "Basis Backend",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Radiance Image"
      }
    },
...
    "/radiance/manifest": {
      "post": {
        "operationId": "get_radiance_manifest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RadianceRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RadianceManifestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Radiance Manifest"
      }
    },
...
  }
}